introduce `linear_call` for custom transposition.
This adds a primitive with a corresponding "traceable" function in
`custom_derivatives` that takes a callee and its transpose, both
functions. When the corresponding primitive is encountered during
transposition, its custom transpose function is invoked instead of
transpose-transforming the callee. The invocation of the custom
transposition is itself done via a `linear_call`, with the original
callee set as the transpose. This maintains, in particular, that
transposing twice is an identity.